requirements.txt
flask-cors==5.0.0
google-generativeai==0.8.3
grpcio-status==1.69.0
pip-chill==1.0.3
python-dotenv==1.0.1
